found the problem but...
i have no clue how im supposed to fix this. the issue is with log(r) just making the entire shader die for some reason, but literally every other function except for the log and log2 function work.

the weird bit is that i'm fairly sure it has nothing to do with the log part of the log function... if i do log of any constant, it works. but when i do a log of a VARIABLE, the thing dies. WHAT AM I SUPPOSED TO DO