... we have a serious problem
so as it turns out, you can't create arrays in Godot shaders with dynamic size... it's one size on declaration and that's it.

this is an issue because i won't be able to change the max period size via a shader parameter, which is kind of important i feel.

dang it.